Fuzzy matching, halve punten en CSV import toegevoegd
- quiz.php: similar_text() voor bijna-goed feedback (+0.5 punt bij >=75% overeenkomst)
- quiz.php: oranje feedbackstijl voor bijna-correcte antwoorden
- resultaat.php: bijna-goede woorden gemarkeerd met ≈ +½
- admin/importeer.php: CSV import met preview, auto-detect separator, nieuw of bestaande lijst
- admin/index.php: knop CSV importeren toegevoegd
- database.sql: score kolom gewijzigd naar DECIMAL(4,1)

+++ docs/database.sql
@@ -47,7 +47,7 @@ CREATE TABLE resultaten (
     id INT AUTO_INCREMENT PRIMARY KEY,
     user_id INT NOT NULL,
     woordenlijst_id INT NOT NULL,
-    score INT NOT NULL,
+    score DECIMAL(4,1) NOT NULL,
     totaal INT NOT NULL,
     created_at DATETIME DEFAULT CURRENT_TIMESTAMP,
     FOREIGN KEY (user_id) REFERENCES users(id) ON DELETE CASCADE,
